data-processing-service/pyproject.toml
[tool.poetry]
name = "data-processing-service"
version = "0.1.0"
description = ""
authors = ["Michael Marinez <80479622+MixedMachine@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "data_processing_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
nats-py = "^2.3.1"
pymongo = "^4.5.0"
sqlalchemy = "^2.0.19"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
